Venomous.SerpentWatcher (Venomous v0.4.1)

🐍🔎 📁 A snake spy watching every step...

Disabled by default. config :venomous, :serpent_watcher, enable: true to enable.

Starts python watchdog process, watching over every python module inside module_paths directories. Default watcher requires watchdog module, which can be installed with mix venomous.watchdog install Whenever file is edited/created/deleted watcher casts :reload with module name to SnakeManager. The default reload module function imports and reloads the module from its name.

## Configuration

config :venomous,
  serpent_watcher: [
    enable: true,
    module: :serpent_watcher, # default
    func: :watch_directories, #default
    args: [Venomous.SnakeManager] # default
  ]

config :venomous, :snake_manager, %{
  ...
  reload_module: :reload, # default. reload function is hard coded to :reload
  ...
}

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

handle_continue(atom, state)

Callback implementation for GenServer.handle_continue/2.

init(args)

Callback implementation for GenServer.init/1.

start_link(args)