Venomous.SnakeArgs (Venomous v0.4.1)

A module to handle arguments for a Python function call.

This module defines a struct for storing the module name, function name, and the list of arguments that can be used to call a Python function.

Types

t()

@type t() :: %Venomous.SnakeArgs{args: [any()], func: atom(), module: atom()}

Functions

from_params(module, func, args)

@spec from_params(atom(), atom(), [any()]) :: %Venomous.SnakeArgs{
  args: term(),
  func: term(),
  module: term()
}

Builds SnakeArgs struct from params

Parameters

module atom() of python module ex. :builtins

function atom() of function from given module ex. :sum

args list(any()) list of arguments for function ex. [ [1,2,3,4,5] ]

Returns

%SnakeArgs{}