Venomous (Venomous v0.3.0)

Venomous is a wrapper around erlport python Ports, designed to simplify concurrent use. It focuses dynamic extensibility, like spawning, reusing and killing processes on demand. Furthermore, unused processes get automatically killed by scheduled process which can be configured inside config.exs. Venomous core functions capture and handle :EXIT calls ensuring that all python process die with it and do not continue their execution.

The core concept revolves around "Snakes" which represent Python worker processes. These Venomous.SnakeWorker are managed and supervised with Venomous.SnakeManager GenServer to allow concurrent and efficient execution of Python code. The Snakes pids and python pids are stored inside :ets table and the Processes are handled by DymanicSupervisor called Venomous.SnakeSupervisor. The unused Snakes get automatically killed by SnakeManager depending on the given configuration.

Main Functionality

python/2: The primary function to execute a Python function. It retrieves a Snake (Python worker process) and runs the specified Python function using the arguments provided in a SnakeArgs struct. If no ready Snakes are available, a new one is spawned. If max_children is reached it will return an error with appropriate message.

python!/2 | python!/1: Will wait until any Venomous.SnakeWorker is freed, requesting it with the given interval.

Architecture

Venomous consists of several key components:

Venomous.SnakeWorker: Manages the execution of Python processes.

Venomous.SnakeSupervisor: A DynamicSupervisor that oversees the SnakeWorkers.

Venomous.SnakeManager: A GenServer that coordinates the SnakeWorkers and handles operations like spawning, retrieval and cleanup.

Configuration Options

SnakeManager

The behavior and management of Snakes can be configured through the following options inside :venomous :snake_manager config key:

snake_ttl_minutes: non_neg_integer(): Time-to-live for a Snake in minutes. Default is 15 min.

perpetual_workers: non_neg_integer(): Number of Snakes to keep alive perpetually. Default is 10.

cleaner_interval: non_neg_integer(): Interval in milliseconds for cleaning up inactive Snakes. Default is 60_000 ms.

erlport_encoder: %{module: atom(), func: atom(), args: list(any())}: Optional :erlport encoder/decoder python function for converting types. This function is applied to every unnamed python process started by SnakeManager. For more information see Handling Erlport API

Python options

Python options can be configured inside :venomous :python_opts config key All of these are optional. However you will most likely want to set module_paths

  @available_opts [
  :module_paths, # List of paths to your python modules
  :cd, # Change python's directory on spawn. Default is $PWD
  :compressed, # Can be set from 0-9. May affect performance. Read more on [Erlport documentation](http://erlport.org/docs/python.html#erlang-api)
  :envvars, # additional python process envvars
  :packet_bytes, # Size of erlport python packet. Default: 4 = max 4GB of data. Can be set to 1 = 256 bytes or 2 = ? bytes if you are sure you won't be transfering a lot of data.
  :python_executable # path to python executable to use.
]

## Auxiliary Functions

- list_alive_snakes/0: Returns a list of :ets table containing currently alive Snakes. - clean_inactive_snakes/0: Manually clears inactive Snakes depending on their ttl and returns the number of Snakes cleared. - slay_python_worker/2: Kills a specified Python worker process and its SnakeWorker. :brutal can be specified as option, which will kill -9 the os process of python which prevents the code from executing until it finalizes or goes through iteration. - retrieve_snake/0: Retrieves a Venomous.SnakeWorker and sets its status to :retrieved - get_snakes_ready/1: Retrieves given amount of Venomous.SnakeWorkers

Functions

clean_inactive_snakes()

@spec clean_inactive_snakes() :: non_neg_integer()

Clears inactive snakes manually, returns number of snakes cleared.

get_snakes_ready(amount)

@spec get_snakes_ready(non_neg_integer()) :: [Venomous.SnakeWorker.t()]

Retrieves x amount of ready snakes and sets their status to :retrieved. In case of hitting max_children cap, stops and returns all available snakes.

Warning

In case of retrieving all available snakes and not using them right away, functions like python!/2 and retrieve_snake!/0 will loop until they are freed.

Parameters

amount of snakes to retrieve

Returns

A list of Venomous.SnakeWorker structs

list_alive_snakes()

@spec list_alive_snakes() :: [{pid(), pid(), non_neg_integer(), atom(), any()}]

Returns list of :ets table containing alive snakes

python(snake_args, opts \\ [])

@spec python(
  Venomous.SnakeArgs.t(),
  keyword()
) :: any()

Wrapper for calling python process Tries to retrieve Venomous.SnakeWorker which then runs the given Venomous.SnakeArgs. In case of failure will return {:retrieve_error, message}. In case :EXIT happens, it will kill python os process along with its worker and exit(reason)

Parameters

Venomous.SnakeArgs struct of :module, :func, :args

opts \ []

Opts

:python_timeout ms timeout. Kills python OS process on timeout. Default: 15_000

:kill_python_on_exception Should python process be killed on exception. Should be set to true if your python process exits by itself. Default: false

Returns

any() | %{error: :timeout} | {retrieve_error: any()} retrieves output of python function or error

python!(snake_args, opts \\ [])

@spec python!(
  Venomous.SnakeArgs.t(),
  keyword()
) :: any()

If no Snake is available will continue requesting it with the given interval until any gets freed or receives :EXIT signal

Opts

:retrieve_interval ms to wait before requesting snake again Default: 200

:python_timeout ms timeout. Kills python OS process on timeout. Default: 15_000

:kill_python_on_exception Should python process be killed on exception. Should be set to true if your python process exits by itself. Default: false

retrieve_snake()

@spec retrieve_snake() ::
  {:retrieve_error, reason :: term()} | Venomous.SnakeWorker.t()

Retrieves Venomous.SnakeWorker struct and sets it's status to :retrieved preventing other processes from accessing it. If all processes are busy and exceeds max_children will return {:retrieve_error, message}.

Returns

Venomous.SnakeWorker struct. In case of error {:retrieve_error, message}

retrieve_snake!(interval \\ 100)

@spec retrieve_snake!(non_neg_integer()) :: Venomous.SnakeWorker.t()

If all processes are busy and exceeds max_children will wait for interval ms and try again. Traps the exit signals, to safely escape loop.

Parameters

interval: The time to wait in milliseconds before retrying. Default is @wait_for_snake_interval.

Returns

Venomous.SnakeWorker struct.

slay_python_worker(snake_worker, termination_style \\ :peaceful)

@spec slay_python_worker(
  Venomous.SnakeWorker.t(),
  termination_style :: atom() | Venomous.SnakeWorker.t()
) :: :ok

Kills python process and its SnakeWorker :brutal also kills the OS process of python, ensuring the process does not continue execution.

Parameters

Venomous.SnakeWorker struct

a Way to kill process. :brutal additionally kills with kill -9 ensuring the python does not execute further. Default: :peaceful

Returns

:ok

snake_run(snake_args, worker, opts \\ [])

@spec snake_run(Venomous.SnakeArgs.t(), Venomous.SnakeWorker.t(), keyword()) :: any()

Runs Venomous.SnakeArgs inside given Venomous.SnakeWorker. Traps exit and awaits signals [:SNAKE_DONE, :SNAKE_ERROR, :EXIT] In case of an exit, brutally kills the python process ensuring it doesn't get executed any further.

Parameters

Venomous.SnakeArgs struct of :module, :func, :args

Venomous.SnakeWorker struct

opts Keywords

Opts

:python_timeout ms timeout. Kills python OS process on timeout. Default: 15_000

:kill_python_on_exception Should python process be killed on exception. Should be set to true if your python process exits by itself. Default: false

Returns

any() | %{error: :timeout} | %SnakeError{} retrieves output of python function or error